Make config_dir modifiable via $COMPETITIVE_VERIFY_CONFIG_PATH env var

+++ src/competitive_verifier/oj.py
@@ -6,15 +6,13 @@ from contextlib import nullcontext
 from logging import getLogger
 from typing import Optional
 
-import onlinejudge._implementation.utils
-import onlinejudge.utils
 import onlinejudge_command.main
 from onlinejudge.service.atcoder import AtCoderService
 from onlinejudge.service.library_checker import LibraryCheckerProblem
 from onlinejudge.service.yukicoder import YukicoderService
 from onlinejudge.type import NotLoggedInError
 
-import competitive_verifier.config
+import competitive_verifier.config as config
 import competitive_verifier.exec
 import competitive_verifier.oj_download_command
 from competitive_verifier import log
@@ -22,20 +20,17 @@ from competitive_verifier.models.result import TestcaseResult, VerificationResul
 from competitive_verifier.models.result_status import ResultStatus
 from competitive_verifier.oj_test_command import run as run_test
 
-_oj_cache_dir = competitive_verifier.config.cache_dir.resolve() / "online-judge-tools"
-_problem_cache_dir = competitive_verifier.config.cache_dir / "problems"
-onlinejudge._implementation.utils.user_cache_dir = _oj_cache_dir
 logger = getLogger(__name__)
 
 checker_exe_path = "checker.exe" if sys.platform == "win32" else "checker"
 
 
 def get_cache_directory() -> pathlib.Path:
-    return _oj_cache_dir
+    return config.get_cache_dir().resolve() / "online-judge-tools"
 
 
 def get_problem_cache_dir() -> pathlib.Path:
-    return _problem_cache_dir
+    return config.get_cache_dir() / "problems"
 
 
 def get_directory(url: str) -> pathlib.Path:
@@ -85,9 +80,8 @@ def download(url: str, *, group_log: bool = False) -> bool:
                     directory=test_directory,
                     cookie=get_cache_directory() / "cookie.txt",
                 )
-
                 checker_path = get_checker_path(url)
-                if checker_path:
+                if checker_path and checker_path.exists():
                     try:
                         shutil.copy2(checker_path, directory)
                     except Exception as e:
@@ -100,7 +94,7 @@ def download(url: str, *, group_log: bool = False) -> bool:
                 elif isinstance(e, NotLoggedInError) and is_atcoder(url):
                     logger.error("Required: $DROPBOX_TOKEN environment variable")
                 else:
-                    logger.exception("Failed to download", e)
+                    logger.exception("Failed to download: %s", e)
                 return False
     else:
         logger.info("download:already exists: %s", url)
